+++ srcs/graphic/multi_sprite.c
@@ -6,7 +6,7 @@
 /*   By: hsabouri <hsabouri@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/03/19 14:29:01 by hsabouri          #+#    #+#             */
-/*   Updated: 2019/05/31 14:36:21 by hsabouri         ###   ########.fr       */
+/*   Updated: 2019/06/16 17:59:44 by hsabouri         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,8 +15,7 @@
 
 t_mat	*choose_entity_material(t_entity entity, t_vec2 pos)
 {
-	float	cosin;
-	float	sinus;
+	t_vec2	cossin;
 	t_vec2	look;
 	t_vec2	perp;
 	t_mat	*mat;
@@ -25,12 +24,12 @@ t_mat	*choose_entity_material(t_entity entity, t_vec2 pos)
 	pos = vec2_normalize(vec2_sub(vec3_to_vec2(entity.physic.pos), pos));
 	look = vec2_rot(vec2_new(0, 1), entity.physic.look_h);
 	perp = vec2_rot(vec2_new(1, 0), entity.physic.look_h);
-	cosin = vec2_cross(pos, look).z;
-	sinus = vec2_cross(pos, perp).z;
-	if (sinus >= 0)
-		img = (entity.mat->len / 4) + cosin * (entity.mat->len / 4);
+	cossin.u = vec2_cross(pos, look).z;
+	cossin.v = vec2_cross(pos, perp).z;
+	if (cossin.v >= 0)
+		img = (entity.mat->len / 4) + cossin.u * (entity.mat->len / 4);
 	else
-		img = (3 * entity.mat->len / 4) - cosin * (entity.mat->len / 4);
+		img = (3 * entity.mat->len / 4) - cossin.u * (entity.mat->len / 4);
 	mat = *(t_mat **)anth(entity.mat, img % entity.mat->len);
 	return (mat);
 }
